move Database and include env file
