+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules/
-.env
 coverage/
 *.log
 npm-debug.log*
