+++ Backend/.env
@@ -1,0 +1,5 @@
+NODE_ENV=development
+PORT=5000
+MONGO_URI=mongodb+srv://davidtorres:12345david@cluster0.rbvit.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
+JWT_SECRET=fF$8d!pK@2vXz#M9qL^rT*B7sNwY&G0h
+JWT_EXPIRE=30d
