+++ Backend/Database/scripts/initRootUser.js
